removed bugs and slowed down changes
Removed the bugs where the program wouldn’t stop if it reached the top
or bottom of the image, this was because I was using statements like:
if (x or y)<0 or (x and y) >512. This caused the program to only
evaluate the first condition(the x conditions) and completely ignore
the second conditions(y). I also made the dots/crosses change colors
more slowly as well as make the default start position be the center of
the image.